Auth refactor

+++ pybossa/view/applications.py
@@ -37,8 +37,10 @@ from pybossa.core import (uploader, signer, sentinel, json_exporter,
     csv_exporter, importer, flickr)
 from pybossa.model.app import App
 from pybossa.model.task import Task
+from pybossa.model.auditlog import Auditlog
+from pybossa.model.blogpost import Blogpost
 from pybossa.util import Pagination, admin_required, get_user_id_or_ip
-from pybossa.auth import require
+from pybossa.auth import ensure_authorized_to
 from pybossa.cache import apps as cached_apps
 from pybossa.cache import categories as cached_cat
 from pybossa.cache import project_stats as stats
@@ -186,7 +188,7 @@ def app_cat_index(category, page):
 @blueprint.route('/new', methods=['GET', 'POST'])
 @login_required
 def new():
-    require.app.create()
+    ensure_authorized_to('create', App)
     form = AppForm(request.form)
 
     def respond(errors):
@@ -247,8 +249,8 @@ def task_presenter_editor(short_name):
      overall_progress, last_activity) = app_by_shortname(short_name)
 
     title = app_title(app, "Task Presenter Editor")
-    require.app.read(app)
-    require.app.update(app)
+    ensure_authorized_to('read', app)
+    ensure_authorized_to('update', app)
 
     form = TaskPresenterForm(request.form)
     form.id.data = app.id
@@ -331,8 +333,8 @@ def delete(short_name):
     (app, owner, n_tasks,
     n_task_runs, overall_progress, last_activity) = app_by_shortname(short_name)
     title = app_title(app, "Delete")
-    require.app.read(app)
-    require.app.delete(app)
+    ensure_authorized_to('read', app)
+    ensure_authorized_to('delete', app)
     if request.method == 'GET':
         return render_template('/applications/delete.html',
                                title=title,
@@ -389,8 +391,8 @@ def update(short_name):
         return redirect(url_for('.details',
                                 short_name=new_project.short_name))
 
-    require.app.read(app)
-    require.app.update(app)
+    ensure_authorized_to('read', app)
+    ensure_authorized_to('update', app)
 
     title = app_title(app, "Update")
     if request.method == 'GET':
@@ -458,7 +460,7 @@ def details(short_name):
     (app, owner, n_tasks, n_task_runs,
      overall_progress, last_activity) = app_by_shortname(short_name)
 
-    require.app.read(app)
+    ensure_authorized_to('read', app)
     template = '/applications/app.html'
 
     redirect_to_password = _check_if_redirect_to_password(app)
@@ -488,8 +490,8 @@ def settings(short_name):
      overall_progress, last_activity) = app_by_shortname(short_name)
 
     title = app_title(app, "Settings")
-    require.app.read(app)
-    require.app.update(app)
+    ensure_authorized_to('read', app)
+    ensure_authorized_to('update', app)
     app = add_custom_contrib_button_to(app, get_user_id_or_ip())
     return render_template('/applications/settings.html',
                            app=app,
@@ -521,8 +523,8 @@ def import_task(short_name):
                          n_volunteers=n_volunteers,
                          n_completed_tasks=n_completed_tasks,
                          target='app.import_task')
-    require.app.read(app)
-    require.app.update(app)
+    ensure_authorized_to('read', app)
+    ensure_authorized_to('update', app)
     importer_type = request.form.get('form_name') or request.args.get('type')
     all_importers = importer.get_all_importer_names()
     if importer_type is not None and importer_type not in all_importers:
@@ -591,8 +593,8 @@ def setup_autoimporter(short_name):
                          n_volunteers=n_volunteers,
                          n_completed_tasks=n_completed_tasks,
                          target='app.setup_autoimporter')
-    require.app.read(app)
-    require.app.update(app)
+    ensure_authorized_to('read', app)
+    ensure_authorized_to('update', app)
     importer_type = request.form.get('form_name') or request.args.get('type')
     all_importers = importer.get_autoimporter_names()
     if importer_type is not None and importer_type not in all_importers:
@@ -644,8 +646,8 @@ def delete_autoimporter(short_name):
                          overall_progress=overall_progress,
                          n_volunteers=n_volunteers,
                          n_completed_tasks=n_completed_tasks)
-    require.app.read(app)
-    require.app.update(app)
+    ensure_authorized_to('read', app)
+    ensure_authorized_to('update', app)
     if app.has_autoimporter():
         autoimporter = app.get_autoimporter()
         app.delete_autoimporter()
@@ -683,7 +685,7 @@ def task_presenter(short_name, task_id):
     task = task_repo.get_task(id=task_id)
     if task is None:
         raise abort(404)
-    require.app.read(app)
+    ensure_authorized_to('read', app)
     redirect_to_password = _check_if_redirect_to_password(app)
     if redirect_to_password:
         return redirect_to_password
@@ -741,7 +743,7 @@ def presenter(short_name):
     title = app_title(app, "Contribute")
     template_args = {"app": app, "title": title, "owner": owner,
                      "invite_new_volunteers": invite_new_volunteers(app)}
-    require.app.read(app)
+    ensure_authorized_to('read', app)
     redirect_to_password = _check_if_redirect_to_password(app)
     if redirect_to_password:
         return redirect_to_password
@@ -772,7 +774,7 @@ def tutorial(short_name):
      overall_progress, last_activity) = app_by_shortname(short_name)
     title = app_title(app, "Tutorial")
 
-    require.app.read(app)
+    ensure_authorized_to('read', app)
     redirect_to_password = _check_if_redirect_to_password(app)
     if redirect_to_password:
         return redirect_to_password
@@ -787,7 +789,7 @@ def export(short_name, task_id):
     (app, owner, n_tasks, n_task_runs,
      overall_progress, last_activity) = app_by_shortname(short_name)
 
-    require.app.read(app)
+    ensure_authorized_to('read', app)
     redirect_to_password = _check_if_redirect_to_password(app)
     if redirect_to_password:
         return redirect_to_password
@@ -808,7 +810,7 @@ def tasks(short_name):
      overall_progress, last_activity) = app_by_shortname(short_name)
     title = app_title(app, "Tasks")
 
-    require.app.read(app)
+    ensure_authorized_to('read', app)
     redirect_to_password = _check_if_redirect_to_password(app)
     if redirect_to_password:
         return redirect_to_password
@@ -854,7 +856,7 @@ def tasks_browse(short_name, page):
                                overall_progress=overall_progress,
                                n_volunteers=n_volunteers,
                                n_completed_tasks=n_completed_tasks)
-    require.app.read(app)
+    ensure_authorized_to('read', app)
     redirect_to_password = _check_if_redirect_to_password(app)
     if redirect_to_password:
         return redirect_to_password
@@ -868,8 +870,8 @@ def delete_tasks(short_name):
     """Delete ALL the tasks for a given project"""
     (app, owner, n_tasks, n_task_runs,
      overall_progress, last_activity) = app_by_shortname(short_name)
-    require.app.read(app)
-    require.app.update(app)
+    ensure_authorized_to('read', app)
+    ensure_authorized_to('update', app)
     if request.method == 'GET':
         title = app_title(app, "Delete")
         n_volunteers = cached_apps.n_volunteers(app.id)
@@ -907,7 +909,7 @@ def export_to(short_name):
     title = app_title(app, gettext("Export"))
     loading_text = gettext("Exporting data..., this may take a while")
 
-    require.app.read(app)
+    ensure_authorized_to('read', app)
     redirect_to_password = _check_if_redirect_to_password(app)
     if redirect_to_password:
         return redirect_to_password
@@ -1121,7 +1123,7 @@ def show_stats(short_name):
     n_completed_tasks = cached_apps.n_completed_tasks(app.id)
     title = app_title(app, "Statistics")
 
-    require.app.read(app)
+    ensure_authorized_to('read', app)
     redirect_to_password = _check_if_redirect_to_password(app)
     if redirect_to_password:
         return redirect_to_password
@@ -1182,8 +1184,8 @@ def task_settings(short_name):
      overall_progress, last_activity) = app_by_shortname(short_name)
     n_volunteers = cached_apps.n_volunteers(app.id)
     n_completed_tasks = cached_apps.n_completed_tasks(app.id)
-    require.app.read(app)
-    require.app.update(app)
+    ensure_authorized_to('read', app)
+    ensure_authorized_to('update', app)
     app = add_custom_contrib_button_to(app, get_user_id_or_ip())
     return render_template('applications/task_settings.html',
                            app=app,
@@ -1201,8 +1203,8 @@ def task_n_answers(short_name):
      overall_progress, last_activity) = app_by_shortname(short_name)
     title = app_title(app, gettext('Redundancy'))
     form = TaskRedundancyForm()
-    require.app.read(app)
-    require.app.update(app)
+    ensure_authorized_to('read', app)
+    ensure_authorized_to('update', app)
     if request.method == 'GET':
         return render_template('/applications/task_n_answers.html',
                                title=title,
@@ -1240,8 +1242,8 @@ def task_scheduler(short_name):
                                form=form,
                                app=app,
                                owner=owner)
-    require.app.read(app)
-    require.app.update(app)
+    ensure_authorized_to('read', app)
+    ensure_authorized_to('update', app)
 
     if request.method == 'GET':
         if app.info.get('sched'):
@@ -1288,8 +1290,8 @@ def task_priority(short_name):
                                form=form,
                                app=app,
                                owner=owner)
-    require.app.read(app)
-    require.app.update(app)
+    ensure_authorized_to('read', app)
+    ensure_authorized_to('update', app)
 
     if request.method == 'GET':
         return respond()
@@ -1326,7 +1328,7 @@ def show_blogposts(short_name):
      overall_progress, last_activity) = app_by_shortname(short_name)
 
     blogposts = blog_repo.filter_by(app_id=app.id)
-    require.blogpost.read(app_id=app.id)
+    ensure_authorized_to('read', Blogpost, app_id=app.id)
     redirect_to_password = _check_if_redirect_to_password(app)
     if redirect_to_password:
         return redirect_to_password
@@ -1347,7 +1349,7 @@ def show_blogpost(short_name, id):
     blogpost = blog_repo.get_by(id=id, app_id=app.id)
     if blogpost is None:
         raise abort(404)
-    require.blogpost.read(blogpost)
+    ensure_authorized_to('read', blogpost)
     redirect_to_password = _check_if_redirect_to_password(app)
     if redirect_to_password:
         return redirect_to_password
@@ -1388,18 +1390,18 @@ def new_blogpost(short_name):
     del form.id
 
     if request.method != 'POST':
-        require.blogpost.create(app_id=app.id)
+        ensure_authorized_to('create', Blogpost, app_id=app.id)
         return respond()
 
     if not form.validate():
         flash(gettext('Please correct the errors'), 'error')
         return respond()
 
-    blogpost = model.blogpost.Blogpost(title=form.title.data,
-                                body=form.body.data,
-                                user_id=current_user.id,
-                                app_id=app.id)
-    require.blogpost.create(blogpost)
+    blogpost = Blogpost(title=form.title.data,
+                        body=form.body.data,
+                        user_id=current_user.id,
+                        app_id=app.id)
+    ensure_authorized_to('create', blogpost)
     blog_repo.save(blogpost)
     cached_apps.delete_app(short_name)
 
@@ -1432,7 +1434,7 @@ def update_blogpost(short_name, id):
     form = BlogpostForm()
 
     if request.method != 'POST':
-        require.blogpost.update(blogpost)
+        ensure_authorized_to('update', blogpost)
         form = BlogpostForm(obj=blogpost)
         return respond()
 
@@ -1440,12 +1442,12 @@ def update_blogpost(short_name, id):
         flash(gettext('Please correct the errors'), 'error')
         return respond()
 
-    require.blogpost.update(blogpost)
-    blogpost = model.blogpost.Blogpost(id=form.id.data,
-                                title=form.title.data,
-                                body=form.body.data,
-                                user_id=current_user.id,
-                                app_id=app.id)
+    ensure_authorized_to('update', blogpost)
+    blogpost = Blogpost(id=form.id.data,
+                        title=form.title.data,
+                        body=form.body.data,
+                        user_id=current_user.id,
+                        app_id=app.id)
     blog_repo.update(blogpost)
     cached_apps.delete_app(short_name)
 
@@ -1463,7 +1465,7 @@ def delete_blogpost(short_name, id):
     if blogpost is None:
         raise abort(404)
 
-    require.blogpost.delete(blogpost)
+    ensure_authorized_to('delete', blogpost)
     blog_repo.delete(blogpost)
     cached_apps.delete_app(short_name)
     flash('<i class="icon-ok"></i> ' + 'Blog post deleted!', 'success')
@@ -1485,7 +1487,7 @@ def auditlog(short_name):
      overall_progress, last_activity) = app_by_shortname(short_name)
 
     logs = auditlogger.get_project_logs(app.id)
-    require.auditlog.read(_app_id=app.id)
+    ensure_authorized_to('read', Auditlog, app_id=app.id)
     redirect_to_password = _check_if_redirect_to_password(app)
     if redirect_to_password:
         return redirect_to_password
